добавила функцию постановки/удаления лайка

+++ src/utils/Api.js
@@ -50,21 +50,37 @@ class Api {
       .then(this._checkResponse)
   }
 
-  deleteLike(id) {
-    return fetch(`${this._baseUrl}/cards/${id}/likes`, {
-      method: "DELETE",
-      headers: this._headers,
-    })
-      .then(this._checkResponse)
+  changeLikeCardStatus(id, isLiked) {
+    if(!isLiked) {
+      return fetch(`${this._baseUrl}/cards/${id}/likes`, {
+        method: "PUT",
+        headers: this._headers,
+      })
+        .then(this._checkResponse)
+    } else {
+      return fetch(`${this._baseUrl}/cards/${id}/likes`, {
+        method: "DELETE",
+        headers: this._headers,
+      })
+        .then(this._checkResponse)
+    }
   }
 
-  putLike(id) {
-    return fetch(`${this._baseUrl}/cards/${id}/likes`, {
-      method: "PUT",
-      headers: this._headers,
-    })
-      .then(this._checkResponse)
-  }
+ // deleteLike(id) {
+ //   return fetch(`${this._baseUrl}/cards/${id}/likes`, {
+ //     method: "DELETE",
+ //     headers: this._headers,
+ //   })
+ //     .then(this._checkResponse)
+ // }
+
+ // putLike(id) {
+ //   return fetch(`${this._baseUrl}/cards/${id}/likes`, {
+  //    method: "PUT",
+ //     headers: this._headers,
+ //   })
+ //     .then(this._checkResponse)
+ // }
 
   changeAvatar(avatar) {
     return fetch(`${this._baseUrl}/users/me/avatar`, {
